fix(web): harden personalization persistence
