+++ apps/web/lib/import/index.ts
@@ -1,19 +1,21 @@
 import "server-only";
-import { sql } from "drizzle-orm";
+import { eq, sql } from "drizzle-orm";
 import { db } from "@/lib/db/client";
-import { chats, messages } from "@/lib/db/schema";
 import {
-  listMemories,
-  setMemory,
-  updatePersonalization,
+  assertPersonalizationCapacity,
+  MemoryCapacityError,
 } from "@/lib/db/personalization";
-import { memorySystemPrompt } from "@/lib/personalization/prompt";
 import {
-  MEMORY_CONTEXT_CHAR_LIMIT,
-  MEMORY_ENTRY_LIMIT,
+  chats,
+  memories,
+  messages,
+  userPersonalization,
+} from "@/lib/db/schema";
+import {
   MemoryInputSchema,
   PersonalizationInputSchema,
   type MemoryInput,
+  type Personalization,
   type PersonalizationInput,
 } from "@/lib/personalization/schema";
 import { extractSearchText } from "@/lib/search/extract";
@@ -63,51 +65,52 @@ function runAdapter(format: ImportFormat, data: unknown): ImportedChat[] {
   }
 }
 
+type ImportTransaction = Parameters<Parameters<typeof db.transaction>[0]>[0];
+
 function writeChats(
+  tx: ImportTransaction,
   userId: string,
   imported: ImportedChat[],
 ): { chats: number; messages: number } {
   let chatCount = 0;
   let msgCount = 0;
 
-  db.transaction((tx) => {
-    for (const chat of imported) {
-      const chatId = crypto.randomUUID();
-      tx.insert(chats)
+  for (const chat of imported) {
+    const chatId = crypto.randomUUID();
+    tx.insert(chats)
+      .values({
+        id: chatId,
+        userId,
+        title: chat.title.slice(0, 200),
+        createdAt: chat.createdAt,
+        updatedAt: chat.createdAt,
+      })
+      .run();
+
+    for (const msg of chat.messages) {
+      const msgId = crypto.randomUUID();
+      tx.insert(messages)
         .values({
-          id: chatId,
-          userId,
-          title: chat.title.slice(0, 200),
-          createdAt: chat.createdAt,
-          updatedAt: chat.createdAt,
+          id: msgId,
+          chatId,
+          role: msg.role,
+          parts: msg.parts,
+          metadata: msg.metadata,
+          createdAt: msg.createdAt,
         })
         .run();
 
-      for (const msg of chat.messages) {
-        const msgId = crypto.randomUUID();
-        tx.insert(messages)
-          .values({
-            id: msgId,
-            chatId,
-            role: msg.role,
-            parts: msg.parts,
-            metadata: msg.metadata,
-            createdAt: msg.createdAt,
-          })
-          .run();
-
-        const content = extractSearchText(msg.parts);
-        if (content) {
-          tx.run(sql`
-            INSERT INTO messages_fts (content, message_id, chat_id, user_id)
-            VALUES (${content}, ${msgId}, ${chatId}, ${userId})
-          `);
-        }
-        msgCount++;
+      const content = extractSearchText(msg.parts);
+      if (content) {
+        tx.run(sql`
+          INSERT INTO messages_fts (content, message_id, chat_id, user_id)
+          VALUES (${content}, ${msgId}, ${chatId}, ${userId})
+        `);
       }
-      chatCount++;
+      msgCount++;
     }
-  });
+    chatCount++;
+  }
 
   return { chats: chatCount, messages: msgCount };
 }
@@ -143,28 +146,67 @@ function readNativePersonalization(data: unknown): NativePersonalization {
   return { personalization, memories: imported };
 }
 
-async function writeNativePersonalization(
+function writeNativePersonalization(
+  tx: ImportTransaction,
   userId: string,
   imported: NativePersonalization,
-): Promise<number> {
+): number {
   if (!imported.personalization && imported.memories.length === 0) return 0;
-  const current = await listMemories(userId);
+  const current = tx
+    .select({ key: memories.key, value: memories.value })
+    .from(memories)
+    .where(eq(memories.userId, userId))
+    .all();
   const merged = new Map(
     current.map((memory) => [memory.key, { key: memory.key, value: memory.value }]),
   );
   for (const memory of imported.memories) merged.set(memory.key, memory);
-  const prospective = [...merged.values()];
-  if (
-    prospective.length > MEMORY_ENTRY_LIMIT ||
-    (memorySystemPrompt(prospective)?.length ?? 0) > MEMORY_CONTEXT_CHAR_LIMIT
-  ) {
-    throw new ImportError("Imported memories exceed OvertChat's memory limit.");
+
+  const storedPersonalization = tx
+    .select({
+      enabled: userPersonalization.enabled,
+      preferredName: userPersonalization.preferredName,
+      occupation: userPersonalization.occupation,
+      about: userPersonalization.about,
+    })
+    .from(userPersonalization)
+    .where(eq(userPersonalization.userId, userId))
+    .get();
+  const personalization: Personalization =
+    imported.personalization ??
+    storedPersonalization ?? {
+      enabled: true,
+      preferredName: null,
+      occupation: null,
+      about: null,
+    };
+  try {
+    assertPersonalizationCapacity(personalization, [...merged.values()]);
+  } catch (error) {
+    if (error instanceof MemoryCapacityError) {
+      throw new ImportError(error.message);
+    }
+    throw error;
   }
 
   if (imported.personalization) {
-    await updatePersonalization(userId, imported.personalization);
+    tx.insert(userPersonalization)
+      .values({ userId, ...imported.personalization })
+      .onConflictDoUpdate({
+        target: userPersonalization.userId,
+        set: { ...imported.personalization, updatedAt: new Date() },
+      })
+      .run();
   }
-  for (const memory of imported.memories) setMemory(userId, memory);
+  for (const memory of imported.memories) {
+    tx.insert(memories)
+      .values({ id: crypto.randomUUID(), userId, ...memory })
+      .onConflictDoUpdate({
+        target: [memories.userId, memories.key],
+        set: { value: memory.value, updatedAt: new Date() },
+      })
+      .run();
+  }
   return imported.memories.length;
 }
 
@@ -182,15 +224,18 @@ export async function importChats(
   const normalized = runAdapter(format, data);
   const personalization =
     format === "ours" ? readNativePersonalization(data) : { memories: [] };
-  const counts = writeChats(userId, normalized);
-  const importedMemories = await writeNativePersonalization(
-    userId,
-    personalization,
-  );
-  return {
-    format,
-    importedChats: counts.chats,
-    importedMessages: counts.messages,
-    importedMemories,
-  };
+  return db.transaction((tx) => {
+    const counts = writeChats(tx, userId, normalized);
+    const importedMemories = writeNativePersonalization(
+      tx,
+      userId,
+      personalization,
+    );
+    return {
+      format,
+      importedChats: counts.chats,
+      importedMessages: counts.messages,
+      importedMemories,
+    };
+  });
 }
