+++ apps/web/lib/personalization/prompt.test.ts
@@ -1,5 +1,10 @@
 import { describe, expect, it } from "vitest";
-import { memorySystemPrompt, userProfileSystemPrompt } from "./prompt";
+import {
+  memorySystemPrompt,
+  personalizationContextUsage,
+  personalizationSystemPrompt,
+  userProfileSystemPrompt,
+} from "./prompt";
 
 describe("personalization prompt", () => {
   it("omits empty profile and memory sections", () => {
@@ -41,6 +46,26 @@ describe("personalization prompt", () => {
         "- `response_style`: Prefer concise answers.",
         "- `timezone`: Uses America/Los_Angeles.",
       ].join("\n"),
+    );
+  });
+
+  it("measures the complete personalization context in UTF-8 bytes", () => {
+    const personalization = {
+      preferredName: "Boomer",
+      occupation: null,
+      about: "你好",
+    };
+    const memories = [{ key: "style", value: "Concise." }];
+    const context = personalizationSystemPrompt(personalization, memories);
+
+    expect(context).toBe(
+      [
+        "# User profile\nPreferred name: Boomer\nMore about the user: 你好",
+        "# Existing memory about the user\n- `style`: Concise.",
+      ].join("\n\n"),
+    );
+    expect(personalizationContextUsage(personalization, memories).bytes).toBe(
+      new TextEncoder().encode(context ?? "").byteLength,
     );
   });
 });
